# VC look up preformance optimization