Improve VC lookup performance. 

@hppritcha @sungeunchoi @e-harvey 

Updated vc_lookup criterion test results without performance changes:

tiger$ cat test.out.novcperf
type: MAP	npeers:       2 naddrs:       2 ns: 118.400000
type: MAP	npeers:      32 naddrs:       2 ns: 121.600000
type: MAP	npeers:      32 naddrs:      32 ns: 107.400000
type: MAP	npeers:     512 naddrs:       2 ns: 116.600000
type: MAP	npeers:     512 naddrs:      32 ns: 138.700000
type: MAP	npeers:     512 naddrs:     512 ns: 158.100000
type: MAP	npeers:    8192 naddrs:       2 ns: 108.300000
type: MAP	npeers:    8192 naddrs:      32 ns: 127.300000
type: MAP	npeers:    8192 naddrs:     512 ns: 215.400000
type: MAP	npeers:    8192 naddrs:    8192 ns: 194.500000
type: MAP	npeers:  131072 naddrs:       2 ns: 106.800000
type: MAP	npeers:  131072 naddrs:      32 ns: 221.900000
type: MAP	npeers:  131072 naddrs:     512 ns: 416.400000
type: MAP	npeers:  131072 naddrs:    8192 ns: 543.200000
type: MAP	npeers:  131072 naddrs:  131072 ns: 255.500000
type: TABLE	npeers:       2 naddrs:       2 ns: 20.100000
type: TABLE	npeers:      32 naddrs:       2 ns: 21.900000
type: TABLE	npeers:      32 naddrs:      32 ns: 21.300000
type: TABLE	npeers:     512 naddrs:       2 ns: 19.500000
type: TABLE	npeers:     512 naddrs:      32 ns: 19.600000
type: TABLE	npeers:     512 naddrs:     512 ns: 19.600000
type: TABLE	npeers:    8192 naddrs:       2 ns: 19.400000
type: TABLE	npeers:    8192 naddrs:      32 ns: 20.200000
type: TABLE	npeers:    8192 naddrs:     512 ns: 20.000000
type: TABLE	npeers:    8192 naddrs:    8192 ns: 20.200000
type: TABLE	npeers:  131072 naddrs:       2 ns: 44.600000
type: TABLE	npeers:  131072 naddrs:      32 ns: 104.300000
type: TABLE	npeers:  131072 naddrs:     512 ns: 262.400000
type: TABLE	npeers:  131072 naddrs:    8192 ns: 292.700000
type: TABLE	npeers:  131072 naddrs:  131072 ns: 20.500000

With performance changes:

tiger$ cat test.out.vcperf
type: MAP	npeers:       2 naddrs:       2 ns: 47.800000
type: MAP	npeers:      32 naddrs:       2 ns: 41.100000
type: MAP	npeers:      32 naddrs:      32 ns: 42.100000
type: MAP	npeers:     512 naddrs:       2 ns: 37.800000
type: MAP	npeers:     512 naddrs:      32 ns: 46.000000
type: MAP	npeers:     512 naddrs:     512 ns: 69.100000
type: MAP	npeers:    8192 naddrs:       2 ns: 34.100000
type: MAP	npeers:    8192 naddrs:      32 ns: 44.300000
type: MAP	npeers:    8192 naddrs:     512 ns: 77.700000
type: MAP	npeers:    8192 naddrs:    8192 ns: 84.200000
type: MAP	npeers:  131072 naddrs:       2 ns: 36.400000
type: MAP	npeers:  131072 naddrs:      32 ns: 66.200000
type: MAP	npeers:  131072 naddrs:     512 ns: 175.500000
type: MAP	npeers:  131072 naddrs:    8192 ns: 236.700000
type: MAP	npeers:  131072 naddrs:  131072 ns: 103.600000
type: TABLE	npeers:       2 naddrs:       2 ns: 25.800000
type: TABLE	npeers:      32 naddrs:       2 ns: 26.900000
type: TABLE	npeers:      32 naddrs:      32 ns: 26.900000
type: TABLE	npeers:     512 naddrs:       2 ns: 25.000000
type: TABLE	npeers:     512 naddrs:      32 ns: 25.000000
type: TABLE	npeers:     512 naddrs:     512 ns: 25.300000
type: TABLE	npeers:    8192 naddrs:       2 ns: 25.100000
type: TABLE	npeers:    8192 naddrs:      32 ns: 27.200000
type: TABLE	npeers:    8192 naddrs:     512 ns: 25.700000
type: TABLE	npeers:    8192 naddrs:    8192 ns: 27.200000
type: TABLE	npeers:  131072 naddrs:       2 ns: 25.000000
type: TABLE	npeers:  131072 naddrs:      32 ns: 30.500000
type: TABLE	npeers:  131072 naddrs:     512 ns: 38.300000
type: TABLE	npeers:  131072 naddrs:    8192 ns: 45.600000
type: TABLE	npeers:  131072 naddrs:  131072 ns: 26.200000
[====] Synthesis: Tested: 1 | Passing: 1 | Failing: 0 | Crashing: 0