@ztiffany 2016-05-23
I have run the following test:

1. Criterion tests
2. cray_runall.sh suite
3. osu_latency+ompi over GNI

Let me know if there are more tests I should do.  I'm a little unsure if I've got all connection setup paths covered.