@sungeunchoi 2016-05-24
You could also run the OSU suite as we do with our jenkins testing with https://github.com/ofi-cray/fab-utils/blob/master/scripts/benchmarks/run_osu.