@jswaro 2016-05-24
How consistent are the results? The change is really nice. Is there much variance in the results?