@sungeunchoi 2016-05-25
Are these VC_MODE bits even needed anymore?  The vector one is not used, and the hashtable one is only used for sanity checking (which you might be able to in a different way).