@ztiffany 2016-05-25
@jswaro results are pretty consistent.

@sungeunchoi @e-harvey I can make the requested changes to the flags.  I don't think they ever served a function, though.  I assume they are for debugging?